+++ inc/problem.class.php
@@ -1317,8 +1317,8 @@ class Problem extends CommonITILObject {
    }
 
 
-
    static function getCommonSelect() {
+      Toolbox::deprecated('Use getCommonCriteria');
 
       $SELECT = "";
       if (count($_SESSION["glpiactiveentities"])>1) {
@@ -1333,6 +1333,7 @@ class Problem extends CommonITILObject {
 
 
    static function getCommonLeftJoin() {
+      Toolbox::deprecated('Use getCommonCriteria');
 
       $FROM = "";
       if (count($_SESSION["glpiactiveentities"])>1) {
@@ -1371,14 +1372,12 @@ class Problem extends CommonITILObject {
          return false;
       }
 
-      $restrict         = '';
-      $order            = '';
+      $restrict         = [];
       $options['reset'] = 'reset';
 
       switch ($item->getType()) {
          case 'User' :
-            $restrict   = "(`glpi_problems_users`.`users_id` = '".$item->getID()."')";
-            $order      = '`glpi_problems`.`date_mod` DESC';
+            $restrict['glpi_problems_users.users_id'] = $item->getID();
 
             $options['criteria'][0]['field']      = 4; // status
             $options['criteria'][0]['searchtype'] = 'equals';
@@ -1398,8 +1397,7 @@ class Problem extends CommonITILObject {
             break;
 
          case 'Supplier' :
-            $restrict   = "(`glpi_problems_suppliers`.`suppliers_id` = '".$item->getID()."')";
-            $order      = '`glpi_problems`.`date_mod` DESC';
+            $restrict['glpi_problems_suppliers.suppliers_id'] = $item->getID();
 
             $options['criteria'][0]['field']      = 6;
             $options['criteria'][0]['searchtype'] = 'equals';
@@ -1422,13 +1420,7 @@ class Problem extends CommonITILObject {
             }
             echo "</td></tr></table>";
 
-            if ($tree) {
-               $restrict = "IN (".implode(',', getSonsOf('glpi_groups', $item->getID())).")";
-            } else {
-               $restrict = "='".$item->getID()."'";
-            }
-            $restrict   = "(`glpi_groups_problems`.`groups_id` $restrict)";
-            $order      = '`glpi_problems`.`date_mod` DESC';
+            $restrict['glpi_groups_problems.groups_id'] = ($tree ? getSonsOf('glpi_groups', $item->getID()) : $item->getID());
 
             $options['criteria'][0]['field']      = 71;
             $options['criteria'][0]['searchtype'] = ($tree ? 'under' : 'equals');
@@ -1437,9 +1429,8 @@ class Problem extends CommonITILObject {
             break;
 
          default :
-            $restrict   = "(`items_id` = '".$item->getID()."'
-                            AND `itemtype` = '".$item->getType()."')";
-            $order      = '`glpi_problems`.`date_mod` DESC';
+            $restrict['items_id'] = $item->getID();
+            $restrict['itemtype'] = $item->getType();
             break;
       }
 
@@ -1462,17 +1453,11 @@ class Problem extends CommonITILObject {
          echo "</div>";
       }
 
-      $query = "SELECT ".self::getCommonSelect()."
-                FROM `glpi_problems`
-                LEFT JOIN `glpi_items_problems`
-                  ON (`glpi_problems`.`id` = `glpi_items_problems`.`problems_id`) ".
-                self::getCommonLeftJoin()."
-                WHERE $restrict ".
-                      getEntitiesRestrictRequest("AND", "glpi_problems")."
-                ORDER BY $order
-                LIMIT ".intval($_SESSION['glpilist_limit']);
-      $result = $DB->query($query);
-      $number = $DB->numrows($result);
+      $criteria = self::getCommonCriteria();
+      $criteria['WHERE'] = $restrict + getEntitiesRestrictCriteria(self::getTable());
+      $criteria['LIMIT'] = (int)$_SESSION['glpilist_limit'];
+      $iterator = $DB->request($criteria);
+      $number = count($iterator);
 
       // Ticket for the item
       echo "<div><table class='tab_cadre_fixe'>";
@@ -1505,7 +1490,7 @@ class Problem extends CommonITILObject {
       if ($number > 0) {
          self::commonListHeader(Search::HTML_OUTPUT);
 
-         while ($data = $DB->fetch_assoc($result)) {
+         while ($data = $iterator->next()) {
             Session::addToNavigateListItems('Problem', $data["id"]);
             self::showShort($data["id"]);
          }
@@ -1520,24 +1505,17 @@ class Problem extends CommonITILObject {
       if (count($linkeditems)) {
          foreach ($linkeditems as $ltype => $tab) {
             foreach ($tab as $lID) {
-               $restrict[] = "(`itemtype` = '$ltype' AND `items_id` = '$lID')";
+               $restrict[] = ['AND' => ['itemtype' => $ltype, 'items_id' => $lID]];
             }
          }
       }
 
       if (count($restrict)) {
-
-         $query = "SELECT ".self::getCommonSelect()."
-                   FROM `glpi_problems`
-                   LEFT JOIN `glpi_items_problems`
-                        ON (`glpi_problems`.`id` = `glpi_items_problems`.`problems_id`) ".
-                   self::getCommonLeftJoin()."
-                   WHERE ".implode(' OR ', $restrict).
-                         getEntitiesRestrictRequest(' AND ', 'glpi_problems') . "
-                   ORDER BY `glpi_problems`.`date_mod` DESC
-                   LIMIT ".intval($_SESSION['glpilist_limit']);
-         $result = $DB->query($query);
-         $number = $DB->numrows($result);
+         $criteria = self::getCommonCriteria();
+         $criteria['WHERE'] = ['OR' => $restrict]
+            + getEntitiesRestrictCriteria(self::getTable());
+         $iterator = $DB->request($criteria);
+         $number = count($iterator);
 
          echo "<div class='spaced'><table class='tab_cadre_fixe'>";
          echo "<tr><th colspan='$colspan'>";
@@ -1547,7 +1525,7 @@ class Problem extends CommonITILObject {
          if ($number > 0) {
             self::commonListHeader(Search::HTML_OUTPUT);
 
-            while ($data = $DB->fetch_assoc($result)) {
+            while ($data = $iterator->next()) {
                // Session::addToNavigateListItems(TRACKING_TYPE,$data["id"]);
                self::showShort($data["id"]);
             }